The values for Reported RQ were not being updated to the final state
Most of this was handled in luxel_plus_error_handling.py

In final_error_checks we had to ensure that we were updating the final
"Reported_RQ" values in "calculation_results".

Also, the memory references for RadQual throughout each of the functions
was removed and a new variable was defined. We don't want to overwrite
the original RQ but want create a new one which carries the "Reported"
value. Therefore, in each case, a new variable (such as new_RadQual) was
defined instead of using RadQual directly